introduced settings navigation helper
